correct "No such property SDG"

+++ Projeto_Ind_ADC/src/main/java/pt/unl/fct/di/adc/firstwebapp/model/ListEventsRequest.java
@@ -30,8 +30,7 @@ public class ListEventsRequest extends AbstractTokenInputRequest<ListEventsReque
 	    private String organizerUsername;
 	    private Integer pageSize;
 	    private String cursor;
-	   // @JsonProperty("SDG")
-	    private List<Integer> sdg;
+	    private List<Long> sdg;
 	    private Boolean isAccessible;
 
 	    public ListEventsInput() {}
@@ -51,8 +50,8 @@ public class ListEventsRequest extends AbstractTokenInputRequest<ListEventsReque
 	    public String getCursor() { return cursor; }
 	    public void setCursor(String cursor) { this.cursor = cursor; }
 	    
-	    public List<Integer> getSDG() { return sdg; }
-	    public void setSDG(List<Integer> sdg) { this.sdg = sdg; }
+	    public List<Long> getSDG() { return sdg; }
+	    public void setSDG(List<Long> sdg) { this.sdg = sdg; }
 	    
 	    public Boolean isAccessible() { return isAccessible; }
 	    public void setisAccessible(Boolean isAccessible) { this.isAccessible = isAccessible; }
